Update README.md
Line Chart comparing the outcomes of theatre category based on its launch date. Stacked Bar chart comparing the outcomes of Parent categories. Box Plot findings of Goal vs Pledge of musicals in Great Britain as Louise is thinking of doing a musical in the future for a campaign budget of $4000.