Add files via upload
Goal vs Pledge Box Plot for the musical campaign in Great Britain. This finding is for Louise's future musical campaign. The box plot shows that half of the campaigns have a goal of less than $2000 which is lower than the third quartile of the pledge amount. So in order for Louise's musical campaign to be successful, her campaign goal has to be below the third quartile of the pledge amount that is $1,496.